+++ packages/web3-eth-abi/src/api/events_api.ts
@@ -1,12 +1,12 @@
 import { AbiError } from 'web3-common';
 import { sha3Raw } from 'web3-utils';
-import { JsonAbiEventFragment } from '../types';
+import { AbiEventFragment } from '../types';
 import { jsonInterfaceMethodToString, isAbiEventFragment } from '../utils';
 
 /**
  * Encodes the event name to its ABI signature, which are the sha3 hash of the event name including input types..
  */
-export const encodeEventSignature = (functionName: string | JsonAbiEventFragment): string => {
+export const encodeEventSignature = (functionName: string | AbiEventFragment): string => {
 	if (typeof functionName !== 'string' && !isAbiEventFragment(functionName)) {
 		throw new AbiError('Invalid parameter value in encodeEventSignature');
 	}