+++ CHANGELOG.md
@@ -345,3 +345,8 @@ Released with 1.0.0-beta.37 code base.
 4. `recover` function's last param is boolean `hashed`, it is used to indicate if data provided is already hashed or not. By default this function will assume data is not hashed.
 5. The `Wallet` no more supports address/number indexing. Have to use `wallet.get` instead.
 6. `Wallet.create` function doesn't accepts `entropy` param
+
+#### web3-eth-contract
+
+1. Event logs does not support types for indexed properties but named properties supported.
+2. Types for overloaded ABI functions are not yet supported.